+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Fixed example in crate documentation.
 
+### Improvements
+
+- Added fakes for `irq_disable`, `irq_enable` and `wfi`.
+
 ## 0.5.0
 
 ### Breaking changes